Introduce editable and read-only views.
When viewing the content of a table, the view is editable, where user
can can modify data of the table (not implemented yet). To make it
possible, we also have to keep track of the primary key of the current
table being edited.

When viewing other contents, such as table schema, indexes (not
implemented yet), the view is read-only.