Change db interface to use sqlite3_exec.
sqlite3_exec is easier to use than sqlite3_stmt objects.
Some features are disabled, e.g. editing a table, since we need
a better way to keep track of the primary key of a table.
We may prefer to use callback mode of the IupMatrix control,
and the code still needs redesigning.